requirements.txt
python-dotenv
flask
tensorflow
werkzeug
numpy
pillow
langchain-groq
groq
gunicorn
